Refactor APIUrl and APIKey into karpatkit.api_services
