+++ src/karpatkit/explorer.py
@@ -6,7 +6,8 @@ from web3 import Web3
 from defabipedia import Blockchain, Chain
 
 from .cache import cache_call
-from .constants import TESTNET_CHAINS, Address, APIKey, APIUrl
+from .constants import TESTNET_CHAINS, Address
+from .api_services import APIKey, APIUrl
 from .node import get_node
 
 TESTNET_HEADER = {
